pyproject.toml
[project]
name = "data-synthesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.1",
    "func-timeout>=4.3.5",
    "hf-transfer>=0.1.9",
    "hydra-core>=1.3.2",
    "langchain>=1.1.0",
    "langchain-cerebras>=0.6.0",
    "langchain-core>=1.1.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=0.3.34",
    "omegaconf>=2.3.0",
    "qwen-vl-utils>=0.0.14",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
    "vllm>=0.11.2",
    "xgboost>=3.1.2",
]
